Add onInit and onPreInit Promises to App
We might still need events on DOM elements though.

Legacy harviewer and harviewer-react seem to attach to the DOM at
different times (maybe because of requirejs async loading for legacy),
so in harviewer-react the default position is we register for the events
too late, i.e. they've already fired.

Need to find a way of using our lifecycle events like the DOM "load"
event, which will notify a listener even when it registers AFTER the
event has fired.