Add "expand" URL parameter handling
This might not be the final solution, as after all PageList components
are initially expanded, if we unexpand a PageList, then click between
tabs, then come back to Preview, the PageLists are all expanded again.

So we need better state management here, rather than overriding existing
state with the URL parameter.