fix: this.tabBodyRefs now passes
NOTE: This test only works because a HAR is loaded (and therefore the
Preview tab is visible/active).

If a HAR is not loaded on startup, then the Home tab will be
visible/active, the Preview tab will NOT have been created, and
therefore any attempts to call methods such as

viewer.getTab("Preview).showStats()
viewer.getTab("Preview).showTimeline()

will fail as viewer.getTab("Preview) will return null.

TODO: We probably need to rework this part of the HV API for the react
way of doing things.